requirements.txt
appdirs==1.4.4
black==20.8b1
bleach==3.2.1
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.1.1
docutils==0.16
idna==2.10
jeepney==0.4.3
keyring==21.4.0
lxml==4.5.2
mypy==0.782
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
pkg-resources==0.0.0
pkginfo==1.5.0.1
pycparser==2.20
pydantic==1.6.1
pygit2==1.3.0
Pygments==2.7.1
pyparsing==2.4.7
PyYAML==5.3.1
readme-renderer==27.0
regex==2020.9.27
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
SecretStorage==3.1.2
six==1.15.0
toml==0.10.1
tqdm==4.50.2
twine==3.2.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
webencodings==0.5.1
